refactor(coordination): board_reminder.rs を 4 ファイルに分割 (SPEC-1974 Phase 10)
- board_reminder.rs (941 行) を board_reminder/{mod,texts,plan,format}.rs へ分割
- texts.rs に const 文字列・caps・windows を集約 (FOR_YOU_MARKER 新設)
- plan.rs に ReminderInputs / ReminderPlan / pure plan_* と plan tests
- format.rs に format_entry_line / entry_targets_self / 関連 helper と format tests
- mod.rs に IO entry points と build_self_match_keys helper (L2)
- 自分宛 marker を [for-you] から >> に変更し reminder 本文との衝突を解消 (M2、FR-043)
- BoardEntry に with_target_owner singular helper を追加 (L1)
- 既存 20 tests を新 module 配置に分散 + 5 件新規 helper test

Verification:
- cargo test -p gwt-core -p gwt: 全 pass
- cargo clippy -- -D warnings: 0 warnings
- cargo fmt --check: 差分なし
