+++ crates/gwt/src/cli/hook/board_reminder/texts.rs
@@ -1,0 +1,79 @@
+//! Reminder text constants and tuning windows for `board-reminder`.
+//!
+//! All user-facing reminder strings live here so that the orchestration
+//! (`mod.rs`), pure plan (`plan.rs`), and entry formatting (`format.rs`)
+//! modules stay focused on logic. SPEC-1974 FR-036 / FR-041 / FR-043 are
+//! the authoritative source for the wording, marker shape, and
+//! reminder-vs-entry separation.
+
+use chrono::Duration;
+
+pub(super) const SESSION_START_CAP: usize = 20;
+pub(super) const USER_PROMPT_DIFF_CAP: usize = 20;
+
+pub(super) fn session_start_window() -> Duration {
+    Duration::hours(24)
+}
+
+pub(super) fn redundancy_window() -> Duration {
+    Duration::minutes(10)
+}
+
+/// Marker prefix attached to entry lines whose `target_owners` match the
+/// current session (SPEC-1974 FR-041, FR-043). The marker MUST stay distinct
+/// from any verbatim substring inside [`USER_PROMPT_REMINDER`] etc. so that
+/// reminder body and entry-line prefix never collide in test assertions.
+pub(super) const FOR_YOU_MARKER: &str = ">> ";
+
+pub(super) const USER_PROMPT_REMINDER: &str = "# Board Post Reminder\n\
+\n\
+Post to the shared Board when you cross a reasoning milestone OR a coordination boundary, \
+so other agents and the user can collaborate without collision.\n\
+\n\
+**Reasoning axes** (the *why* behind your work):\n\
+- Work phase transitions (e.g., implementation -> build check -> PR handoff). Use `--kind status`.\n\
+- Choices between alternatives with the reasoning behind them (e.g., \"A vs B, chose B because ...\"). Use `--kind decision` or `--kind status`.\n\
+- Concerns or hypotheses you are verifying (e.g., \"Hypothesis: failure stems from Y, verifying ...\"). Use `--kind status`.\n\
+\n\
+**Coordination axes** (so others know what you own and what is next):\n\
+- claim — declare ownership of a scope (e.g., \"I claim feature/X migration; others take other ranges\"). Use `--kind claim`.\n\
+- next — coordinate the next step without picking a recipient (e.g., \"phase 1 done, please pick up phase 2\"). Use `--kind next`.\n\
+- blocked — surface a blocker that needs unblocking (e.g., \"waiting on Y, requesting unblock\"). Use `--kind blocked`.\n\
+- handoff — pass concrete work to another agent or the user (e.g., \"completed Y, handing off the PR\"). Use `--kind handoff`.\n\
+- decision — broadcast a confirmed decision (e.g., \"adopting X for the migration\"). Use `--kind decision`.\n\
+\n\
+Add `--target <session-id|branch|agent-id>` (repeatable) when the post is meant for specific agents. \
+Targeted posts are prefixed with a structured marker (currently the `>>` token) at the start of each entry \
+line in the recipient's reminder injection. Omit `--target` for broadcast.\n\
+\n\
+Do NOT post tool-level reports (e.g., \"running gcc\", \"opening file X\", \"ran test Y\"). \
+Anything already visible in the diff or log does not need a Board entry.\n\
+\n\
+Examples:\n\
+  gwtd board post --kind status --body '<your reasoning>'\n\
+  gwtd board post --kind claim --target feature/foo --body 'taking the migration on feature/foo'\n\
+  gwtd board post --kind handoff --body 'phase 1 done, please pick up phase 2'\n";
+
+pub(super) const USER_PROMPT_REMINDER_SHORT: &str = "# Board Post Reminder\n\
+\n\
+You posted to the Board recently. Post again only if a new reasoning milestone \
+(phase change, alternative chosen, concern raised) or a coordination boundary \
+(claim, next, handoff, blocked, decision) has emerged.\n";
+
+// Stop reminders are emitted as `systemMessage` (user-facing) because
+// Claude Code's Stop hook schema does not accept `hookSpecificOutput`.
+// Phrasing is therefore user-oriented rather than agent-oriented.
+pub(super) const STOP_REMINDER: &str = "Board Post Reminder (Stop): the agent is stopping. If you \
+expect a final handoff, prompt the agent to post what it completed to the shared Board \
+with `gwtd board post --kind status` before handing off.";
+
+pub(super) const STOP_REMINDER_SHORT: &str = "Board Post Reminder (Stop): the agent posted to the \
+Board recently; no additional completed-status post is required before stopping.";
+
+pub(super) const INJECTION_HEADER: &str = "# Recent Board updates\n\n\
+The following reasoning posts were made by other Agents since your last Board context. \
+Consider whether any affect your current work phase. This is context, not a directive — \
+you remain autonomous.\n\n";
+
+pub(super) const SESSION_START_HEADER: &str = "# Current Board state\n\n\
+Recent reasoning posts from other Agents (context, not a directive — you remain autonomous):\n\n";
